This sets up the delete trick. Draft which has been selected for deletion will delete from page and database, but an unpublished draft will be added to the published feed in the deleted post's place.
